fix: Add SimpleGraph networkx compatibility layer and fix test dtype
Root cause: networkx is not a project dependency, so SimpleGraph is
always the active graph implementation on CI. SimpleGraph.add_node()
and add_edge() did not accept keyword arguments (position, distance,
fiber_type), and self.graph.nodes was a method instead of a
subscriptable NodeView. This caused 7 test failures.

Changes:
- Redesign SimpleGraph with NodeView class that supports both
  callable iteration (graph.nodes()) and dict-style subscript
  access (graph.nodes[node].get("position"))
- Accept **attrs on add_node() and add_edge() to match nx.Graph API
- Fix test_predict_returns_float32 assertion: model uses float32
  weights by design (memory efficiency), test expected float64.
  numpy 2.x changed dtype promotion (float32 * float stays float32
  instead of upgrading to float64), exposing the mismatch.

+++ tests/test_efficient_models.py
@@ -38,7 +38,7 @@ class TestEfficientQKDPredictor(unittest.TestCase):
     def test_predict_returns_float32(self):
         p = EfficientQKDPredictor(input_dim=3, max_memory_mb=64, enable_pruning=False)
         p.fit(self.X, self.y, epochs=5, learning_rate=0.05, batch_size=8)
-        self.assertEqual(p.predict(self.X).dtype, np.float64)
+        self.assertEqual(p.predict(self.X).dtype, np.float32)
 
     def test_fit_accepts_validation_split(self):
         p = EfficientQKDPredictor(input_dim=3, max_memory_mb=64, enable_pruning=False)
